Day 13 both parts
There are two implementations for part 2 in the library, the earlier one
being left out of interest and to show my shift in strategy as it
happened.

The first implementation was a pretty naive approach: basically try
firewalls with various starting delays, walk them individually, and see
if they get caught. This works, but is comparably very slow since its
iterating the entire state of the firewall as a simulation, and at most
points in time you don't care about most of that state.

The second strategy, which terminated nearly instantly, is much more
efficient. All you really care about is each time `t` when the packet
interesects with each layer with a scanner, and what the position of the
scanner in that layer is at that time. So this approach uses an
algebraic approach to determine the state of a given layer at time `t`,
and then for each layer calculates at what time the packet will reach it
(from a given negative starting position, which is equivalent to a
delay), and whether at that time the scanner is at position 0 So this
approach uses an algebraic approach to determine the state of a given
layer at time `t`, and then for each layer calculates at what time the
packet will reach it (from a given negative starting position, which is
equivalent to a delay), and whether at that time the scanner is at
position 0.