Playing around with racket
Experimenting with racket on last years day1, considering trying to
learn it for this year.

Comments in file with some stumbling blocks.

Not sure if I've got a serious bug or performance for p2 is just awful.
Hasn't terminated yet. Runs fine on small test-case, not real input.
Haven't measured number of loops vs working implementation yet. I had
originally been tracking `seen-freqs` as a list, switched to a vec
thinking it might perform better for the searches. Although I know
allocating another vector on each call is probably not great in racket
either: I don't get the impression it's smart enough to pre-allocate a
big array and then fill it/dynamically grow it.