2019 d3
Kind of brute forced this one. It would have been more efficient to just
store the wire as a vector of the *movements*, not the points, and to
walk the movements as needed. This approach took ~15s on my input, and
I'm sure this could be sub-second, but this is one of those cases where
the brute-force approach was tolerable, and I think it made part 2 much
faster for me to finish.