2019 d7 p1
I had a really stupid bug in my int machine code from the last day with
handling input that I guess I was just really lucky didn't impact that
my answers on day 5. That tied me up for a bit.

Actually that might imply my program for day 5 didn't actually use it's
input? The bug was that I was grabbing the input's destination index on
the tape from the wrong index, so it was getting written to the wrong
spot.