2019 d11 p2 done
My ouput is vertically mirrored, but I was able to read it fine. I guess that
means I was at least consistent in my code for coordinates, but maybe
chose something about my coordinate space different from what was really
expected?

~Also a bit confused about the size of the output: p1 tells me I painted
a little over 2k panels, but in this ouput, I got a range in X of 43 and
range in Y of 6 (that is, max_x - min_x = 42, etc.). That does not fit
2k panels, it fits like 252 panels, right?~

NVM - leaving that above for posterity of my foolishness, but the answer
is easy: the starting condition for p2 is different than p1 because the
starting panel is white. So what p2 paints is totally different than p1,
and the answer from p1 has no bearing on p2.