2019 d18 p1 wip
Playing with Julia. Not fast enough yet.

I didn't really expect this simplistic approach to A* to be good enough,
but I figured I'd give it a shot before doing anything more complex.

I suspect what I should be doing is sort of A* with djikstra to guide
next states. E.g. for each state, look for keys that are actually
reachable from there using djikstra, and then make those the next states
rather than taking 1 step at a time. Trying to capture the entire state
of all the keys at once with a heuristic in A* result in much too broad
a search space & without sufficient focus to drive search in the right
direction quickly.