2020 d14 p2
This seems like an unusual AOC case where the *example* is not
computable with a naive approach, but the *input* (at least my input)
is.

When the mask has lots of wildcards ("X"), fully expanding it is a very
bad idea. But the masks in my input all had a pretty low number of
wildcards (the highest was 8, for 256 distinct expanded values). The p1
example program had 34 wildcards (17 billion expansions), although the
p2 examples were more reasonable.

I think the gold-plated solution here is to not actually expand the
wildcards, but to remember the stored values *with* the wildcard
addresses, and then count out how many expansions there would be at the
end & multiply the values. That's more challenging, though, because
later addresses can intersect with earlier addresses (although maybe in
the given inputs, they don't? I didn't check that), so I think you'd
want to end up storing it as some kind of tree so you track the ordering
and which particular expansions of a given address were overridden
later.  That's an interesting problem to consider, but it's Monday
morning & I have actual work to do, so I'm taking my gold stars where I
can get them.