2020 d15 p2 - brute forcing wins again
Another day where p2 was surprisingly gentle. 30M iterations
isn't *that* many, and I had anticipated something like this & used a
more efficient data structure than just a list of every number spoken in
order.

Brute forcing took 46s to run for me. That's not great, but it's the
right answer! I think there's probably some kind of polynomial that can
express the sequence and would be effectively instant, but it would take
me more than 46s to figure that out, so...