2020 d15 - alternative julia impl
Since today's actual algorithm was pretty simple, but computation-heavy,
I thought it would be a useful comparison point for ruby vs julia
implementations for AOC.

For both parts, the julia impl is ~2.75s. That's a lot better than ruby,
but still slower than I expected! Go or Rust or one of those languages
would probably be the way to go to get sub-second with this algorithm.