2020 d23 p2 wip
Pretty sure this is logically correct, but way too slow. A linked list
was the right call for the manipulations we do, but it's very, very bad
for the search we need to do for the destination cup.

I'll need a secondary data structure to improve that, I think. Maybe I
can keep an array where `arr[x] = node`, where `x` is the label of a cup &
`node` is a reference to the node for that cup? That should make it
quicker - then I can basically just go down through possible destination
values and at each point make sure that destination isn't one of the
picked-up cups.