2015 d04 - parallel implementation
Parallel crunching on hashes was a pretty obvious perf improvment for
this one, and it was the first one of the year where the naive
implementation wasn't nearly instant, so I spent some time learning some
of Nim's concurrency patterns.

Just doing an optimized release build, even without rewriting for
threads, was a pretty considerable ~3x improvement (from 17.9s to 5.0s).
Threading was a further almost 4x improvement (from 5.0s to 1.3s)

I thought about using a channel to communicate the batches of numbers to
each worker, but figured I would still have needed to track the maximum
number passed off to the workers and a lock to wrap access to the
answer, etc., so I decided this was probably simpler in a case like this
where the endpoint is uncertain.

I played around with the batch size a bit - unsurprisingly very small
and very large were both slower (small batches are presumably slower
because of more time spent competing for the lock and large batches are
slower because each worker finishes the current batch before potentially
returning if an answer has been found), but everything in the
~5,000-20,000 range was pretty similar.

It took me a while to actually get this working because nim's approach
to passing references was surprising/confusing to me. I thought
declaring my type as `ref object` meant have a `var x` would get passed
around by reference to methods, but apparently not! At least not when
threads are involved - the `var` was also needed in the `proc`
declaration for `asyncWorker`, where the variable "crosses the boundary"
into threaded code. It does not seem to be needed for other procs where
pass-by-ref is equally important, e.g. `fetchNs`, and in single-threaded
code it wouldn't be needed at all. I don't quite get this, the docs do
describe the memory model around threading as different, and actually
explicitly say you *can't* start a thread with a proc using `var`
arguments, but for me that's how I had to fix it! The docs don't really
demonstrate how you *are* supposed to share heap between threads, but
that's a pretty common need, isn't it? I suppose this means the code
here could break in a future version of nim, but it's good enough for me
for now.

Also realized all my commands in the `Makefile` were invalid while
testing this - oops.