2015 d06 p2
The nim solution here segfaulted. I spent a bit of time trying to figure
out where/why, didn't make much headway, and didn't feel like pulling
out gdb and getting serious about it, so I bailed and rewrote in python.

This might be a bug in nim? The segfault had something to do with the
new `applyCommand` implementation, but `echo`s weren't logging anything,
even with immediate `flushFile(stdout)` to ensure nothing was buffered.
I don't think it's related to uint bounds - I tested some uint
wraparound in the nim playground, and it did wraparound, there was no
bounds checking. That shouldn't be a segfault, anyway.

Interestingly, the p2 implementation in nim works fine if I comment out
the couple lines earlier in `main` to run p1. That makes me more
convinced this is a nim bug.