2015 d08
Love it when there's a clever usage of stdlib that mkaes the whole thing
a couple lines.

Don't use `eval` in production, of course. I trusted AOC not to try and
compromise my computer with a puzzle's input, and I did read the file
first to be sure. Caveat lector, etc.