2015 d19 p2
This probably wasn't the fastest way for me to get to an answer. Some
kind of search was my first instinct, and when it became very clear that
A* searching *forwards* from e -> goal was completely infeasible, I did
some reasearch of other people's solutions and found that searching
backwards was probably faster. So I ended up doing that, and it was ok
with the first cut that did a pretty naive "priority queue" with an
array and a call to `sort_by!` after each loop - about 25s. That would
have been good enough, but I decided to have some fun and see how much a
proper priority queue helped - the basic queue implementation didn't
help much at all because checking membership was expensive. Adding a
set for membership checks to the queue got this down to 7s, which is
good enough for me. I saw some other solutions based on random sampling
that probably would have run faster and been less code, but wouldn't
have been as semantically "correct" or elegant.