2021 d06
First challenging day of the year that left me stumped for a while! It
was fun to iterate through different things that weren't going to work
but helped me understand how to break the intractable bits down into
solvable chunks.

It was immediately clear that p2 couldn't be brute forced. The first
thing I realized is that you don't need to actually simulate the entire
set of fish: if you can figure out how many fish a single fish
with, say, counter `2` ends up as after `d` days, and do the same for
the other starting counters, you can just sum up those numbers for the
final answer. But seeding a single fish and then directly simulating for
hundreds of days is still completely infeasible, so there had to be a
smarter way to figure out how many fish a single fish with counter `c`
turns into after `d` days.

I spent some time trying to figure out if there was a polynomial
equation that could calculate it directly, something in the general
neighborhood of `2 ** (days / 7) + ?` or something, and didn't get
anywhere. I still think there's something to this, but it's beyond my
(rudimentary) mathematical intuition.

Next I started thinking about how to optimize a calculation for a single
fish: simply looping and counting 7 days at a time wasn't going to work,
but made me realize that it was starting to look like a nicely divisible
recursive problem. I.e. if you have `fish = 2` and `days = 80`, at `t =
7` you will have your 1 original fish plus however many fish you have
with `fish = 4, days = 73` (since your original fish spawns another fish
with counter 8 on day 3, and by day 7 the original fish has looped to
its original count and its spawn is down to the original counter + 2).
That seemed solvable, and also seemed efficient to memoize (there aren't
that many pairs of `(fish, days)`). Using my favorite Ruby Hash trick of
a hash-as-a-memoizing-function, the implementation for that recursive
calculation ended up being quite efficient.

Now that I've figured it out, I've looked over some of the solutions on
the subreddit, and I see that the 9-element-array bucket approach is a
more efficient approach that I missed. But the recursive approach is
totally fine, so I'm ok with where I ended up.