2016 d11
A blast from the past I spent some time revisiting last night.

In 2016 I had attempted to do most of the puzzles in Haskell, and this
was one case where p2 tangled me up a lot. My search algorithm in
haskell was not very good, and I had a bug somewhere in "next state"
logic that caused it to get stuck flipping between the same states near
the end, and I just wasn't very capable at figuring out Haskell while
also figuring out those challenges. On top of which it wasn't very fast.

Ruby's not fast either, but in revisiting some old puzzles that I got
stuck on it has the benefit of being more familiar to me (i.e. I have
the skills to debug and tune performance).

This was still a challenging puzzle! The logic of figuring out valid
moves is a little finicky and does require some close-reading of the
puzzle, but doing it in a language I'm initimately familiar with made it
easier to get it done than my attempt in 2016.

The really serious challenge for me on this puzzle continues to be
performance. This implementation still isn't fast, and I feel like I
must be missing an optimization related to finding valid next states or
the cost-to-goal estimation heuristic or something. I haven't actually
sat down to read other's solutions yet, but I did glance at some reddit
posts to sanity check I wasn't wildly off base with my approach, and A*
& BFS seem to be the common choices. My p1 was slow-but-tolerable at
~32s, but p2 took over an hour. It did get the right answer, though!

So I'm not fully satisfied with this, but it is "complete", so I'm
committing it. I may review some other implementations and see if I can
take another crack at making the performance tolerable in the future if
I feel like it.

I left in some code that was really only useful for debugging, e.g.
`to_s` implementations.