2016 d23
This was one of those building-on-a-previously-implemented-vm days, and
I'd done the previous ones in Haskell back in 2016. But I looked, and I
didn't do any extraction for those days (I just copy-pasted), and the
instruction set was tiny, so rewriting seemed faster than getting my
Haskell env set up again.

Part 1 was straightforward. Part 2 was the kind of "manually decode
assembly into a program you can understand/optimize" thing that AOC does
once in a while and I'm no good at. Thankfully, in this case the problem
description gave a super helpful, very blatant hint, and the actual
optimization needed was pretty straightforward. I figured out the early
bit of the program was doing a multiply via 2 loops, introduced a new
`mlt` instruction, and patched the program. I was a bit worried that I
would also need to decode and patch bits of the section down to the
"jmp -16", which seemed more challenging thaks to the `tgl` in the
middle of it, but that turned out not to be needed: after patching the
single multiply it ran nearly instantly.

2016 files were organized a bit differently: I had a `2016/data`
directory, not committed, containing inputs for each day. I'm handling
today a bit differently since I modified the input by hand for part 2,
and because I thought having a version of the assembly with hand
annotations I wrote as I tried to figure it out would be good to keep.