2021 d08
Part 2 definitely twisted my brain a bit. I dead-ended on a few ideas to
try and uniquely map wires by scanning the inputs repeatedly and
eliminating possibilities. Eventually, I realized that after using the
"unique length" digits to start eliminating, I could use that
information to further reduce without the inputs (e.g. if the
unique-length digits show that a => cf, b => cf, d => acf, you can
conclude that d => a and, similarly, nothing else can be a or c or f).
After doing that, every input digit uniquely mapped to a known digit
template even if it still wasn't clear if, say, a => c, b => f or vice
versa.

I'm sure there was a smarter way to actually identify the wire mappings,
and I may have gotten lucky that this worked with my input, but my brain
isn't working great today so this is what we're going with.