2021 d11
I've lost count of the number of times in AOC I made a bug for myself by
building an array of neighboring points and forgetting that a neighbor
of `0` is `-1`, and that's going to actually get an element out of an
array! Probably doesn't help that sometimes I use arrays and sometimes a
hash with keys of points makes more sense.