2018 d20
This one might be my nemesis. This one tangled me up and stumped me for
a really, really long time, both in 2018 and again now.

The appropriate way to handle groups (`(...)`) & options (`|`) escaped
me for a long time, and I eventually got kinda close in 2018 with two
critical mistakes. First, for positions, I was calling `walk` again
rather than just resetting `current_pos`. There's no matching "pop" to
that "push" for options, unlike groups, so that means stuff after the
end of group further down the stack walks from the wrong point.
Secondly, I also wasn't properly "popping" in the group case either and
needed to return when I was just breaking.

The solution I've ended up with is pretty overengineered for the actual
questions. Now that I think I grok this, if I had it do all over again,
I think I could do pretty compact implementation that used djikstra to
walk the regex as-is and build up the mapping of `point -> min distance`
directly, rather than what I ended up doing which was to convert the
regex into a more easily workable "graph" data structure and then apply
a slightly weird version of djikstra to that. The solution I have is,
surprisingly, quite fast (< 500ms for both parts), so I'm not inclined
to do much more here. This is a puzzle I'm just relieved to see the back
of. Also, working through the puzzle this way is how I came to actually
understand it: if I had tried to write djiktra directly against the
regex representation debugging would have been impossible for me.
Spending the extra time building up a graph & code to actually render
the maze from the graph wasn't necessary for getting the answer, but it
was necessary for me to understand how the heck to do this.