2018 d21
Another one that required decompiling the assembly into a more efficient
form to figure out an answer. Either this was easier than other similar
ones or I'm getting better at this process.

The part here that really twisted me up was needing to realize that the
problem description implied the values calculated in r2 (for my input)
were going to loop, and the "lowest integer that executes the most
instructions" would be the last value seen before looping. I may have
missed something crucial, but I actually think both parts focus on
"lowest non-negative integer" is a distraction? My understanding of the
input program is that r0 is effectively a constant set before it starts,
and the program does some weird math in a loop to change the value of r2
until r2 == r0, and r0 is only used for that comparison (it doesn't
influence any calculations). So it doesn't really matter how big or
small r0 is - just that it's a value r2 will eventually equal, and it
matters what order r2 encounters those values in, but that's effectively
a 1:1 mapping. You can't get to different r2 values with the same tick
count, or to the same r2 value with different tick counts (unless you
change r0 after the program has started, I guess, and maybe that's what
the problem description implies you can do but that's getting pretty
complicated).