2021 d18
Time to notice this was a binary tree: a few minutes.

Time to figure out how to walk from a node to the leaf to the left or
right for the "explode" logic: an hour or two.

Time spent tearing my hair out and believing my approach for walking to
the left/right leaf must be flawed, and eventually realizing the bug was
that in some cases there are two nodes right next to each other with the
same values, so with `Struct`'s default `==` being  by-value (a good
default choice and why I often use `Struct` over `class` for these
puzzles), the "walk up until you can turn" bit of the `*_literal_parent`
methods walks too far because it sees the node as equal to the other
side of the node: way, way too much time.