2021 d20 faster
Ok, maybe I was going to fiddle with this a bit more. After I wrote down
the perf change I *wasn't* going to do in the previous commit message,
I decided implementing it would actually be pretty quick, so I did it.

It definitely helps. Runtime for both parts dropped to 27s, a 38%
improvement. Still not very fast, but much better.

Recorded a run with rbspy out of curiosity - looks like the bottleneck
is the "set.include?" checks. I tried some different `Point#hash`
implementations to see if it was a case of bad hash bucket collisions -
a few approaches shaved another couple of seconds off, but nothing very
significant. Changing hash implementations in ruby is a bit of a
low yield optimization most of the time anyway: the default
implementation is a native C function, so the slowdown from a bad hash
has to be pretty bad in order to get a win from replacing that C
function with a ruby function.

The better approach would, I think, just be replacing the set of lit
pixels with a proper matrix/nested arrays. I'd probably want to
pre-allocate them to be large enough for all the pixels I'd want to
track after x rounds of enhancement so I wouldn't need to resize/remap
coordinates after each iteration or anything. Don't think I'm going to
bother trying that, though. I mean it this time.