2021 d23 wip
p1 is working, p2 is currently running.

p1 is correct, but it's all surprisingly slow. 36s for p1, and though
implementation wise p2 is a cake walk, computational-complexity wise
it's a lot worse. p2 seems designed to test efficiency of your
a*/djikstra or whatever, and clearly my impl is already pretty
inefficient here.