dumb bug: don't walk onto other amphipods
d'oh - the paths_from path includes *source* pos but not *dest* pos.
Makes steps easy (`path.count`), but I was only checking that other
amphipods weren't on the *path*, not also on the *dest*. So I was
allowing illegal states where 2 amphipods were on the same spot.

A bit surprising it still found the right answer, but does help explain
why it was slow. This does cut p1 time down to 10s, which is still not
wonderful but is much better than I was doing, and also noticeably
better than the earlier, more naive impl taking ~36s.