adjust path calc to include dest instead of start
keep steps == path.count, but eliminates the need for the extra "dest
occupied by other amphipod check", and makes the `path.any?` check 1.
sufficient and 2. need 1 less check since it was needlessly checking the
start pos before.

This is really nit-picky and splitting hairs at this point and not going
to make a noticeable perf difference, I'm spinning my wheels a bit. p1
is fast enough for me, but p2 is still not good enough.