2021 d23
A day where the basic algorithmic approach was fairly clear up front,
and from there it was a long, slow slog of optimizing the
implementation.

A* to explore the maze was the obvious choice. I initially did it fairly
naively, which each iteration just stepping a single amphipod a single
step and analyzing possible steps every time, and expecting a lot of
little details like some of the rules would come out in the wash (since
some violating some of the rules could never lead to a lower-energy
outcome checking for them would reduce the search space but not change
the answer).

I think that was mostly correct, but it was also very slow. So then I
focused on reducing the search space by encoding the various rules the
problem describes, plus some additional rules that make sense (e.g. when
walking into your own goal room, never stop halfway back if there's
room, and never block an amphipod of another type in your own room).

A lof ot getting this fast was also caching and nitty gritty
implementation choices. A big one was noticing that since the maze is
pretty small, for any given current position there's only a handful of
possible destinations. So I precalculated every possible path with
djikstra, filtering out always-illegal destinations (e.g. blocking a
room or starting/stopping in the hallway).

Two things here are still confusing me a bit, but I'm done trying to
grok them:

- At one point I realized I no longer needed manhattan distance as an
  approximation for `goal_distance` - I had the djikstra paths, I could
  just look up the actual distance. I figured that would be a more
  accurate scoring function for A* and converge to a solution faster.
  It did, but it got the wrong answer on part 1.
- I'm getting the wrong answer on part 2 for the *example*, but my own
  input is correct. So there's clearly cases I've messed up that
  fortunately my input does not have. I got lucky there, I guess.

3.9s for p1, 2:33 for both parts.