2021 d23 refactor
Perf was still bugging me. Decided to try stripping out all the
abstraction and use of structures like sets and just more or less track
the state of the map directly.

And it's faster! Not particularly readable in parts, though.

I guess I'm not shocked: set membership is certainly faster than walking
large arrays, but there is some overhead and it's *not* faster than
directly indexing into an array when you can do that, which I mostly can
here.

p2 is still not particulary fast. p1 is now 2.4s, both parts was 31s.

There's probably a happy middleground where I at least have a `Point`
type. Maybe it would also be faster to dual-track amphipod positions
within the map array and also has a separate hashmap to reduce iterating
through the map. Probably wouldn't really make a difference.

Kind of an interesting final mini-optimization I ended up making:
Rbspy said it was spending 35% of its time in the
`available_paths.reject` block in `next_states_for_amphipod`, and fully
25% specifically checking for other amphipods on the path. I had
constructed a `Set` of "other amphipod positions" at the beginning of
the function and was checking `include?` on that. Changing to just grab
the tile character out of `map` and checking that was faster. Somewhat
surprising to me, checking a regex match was actually slightly faster than
having a `TYPES = ["A", "B", "C", "D"]` constant and checking
`TYPES.include?(x)`. I know the regex matching goes through optimized C
code, but I was still a bit surprised that it edged out checking a 4
element array.

That block in `next_states_for_amphipod` is still the performance
hot-spot. That's not really a *bug* - for a good A* impl searching a
wide search space, I'd expect to spend a lot of time looking for the
next states at the frontier. It just means if I wanted to make this
faster that's probably where I'd focus, but I think I'm done. I still
kind of suspect I do have some kind of larger algorithmic failure here
where I'm considering more states than I need to or not quickly enough
identifying the most promising states, and this still gets the wrong
answer on the p2 example, but it gets the right answers on my input fast
enough for me, I've burnt enough time here, I'm moving on.