2021 d23 refactor
Perf was still bugging me. Decided to try stripping out all the
abstraction and datasets and just more or less track the state of the
map directly.

And it's faster! Not particularly readable in parts, though.

p2 is still not particulary fast. p1 is now 2.9s, both parts was 51s.

There's probably a happy middleground where I at least have a `Point`
type. Maybe it would also be faster to dual-track amphipod positions
within the map array and also has a separate hashmap to reduce iterating
through the map. Probably wouldn't really make a difference, I think the
remaining slowness is just some inefficiencies with generating next
states or scoring them or A* I'm not seeing.