cpp hand-implemented
Given the search space I kind of didn't think even a hand-implemented
brute force approach would work, but I figured I'd try it.

It doesn't work. I mean, it's pretty fast, but there are 10s or 100s of
billions of possible inputs (trillions?), and that's just not realistic
to crunch through.

There's got to be something to do with working backwards from the
constrained ranges in the register values (e.g. often mod or div 26) to
figure out what would validate.