2021 d24
Kind of figured this couldn't be brute-forced, but I started by
implementing a virtual machine anyway to start being able to verify my
interpretation of the program assembly.

Then after reading & figuring out what the program was doing at a
mechanical level, I implemented a more optimized version of that. Given
there's 9^14 valid inputs, I was pretty sure it couldn't be brute forced
that way either, but it was worth trying. But yeah, that was not nearly
fast enough.

From there I basically iterated into a depth first search going
backwards. Basically, since we know what we want the final value of `z`
to be, and we can scan for digits & values of z from the previous step
that result in that. Once you have that, you use that z as the output to
repeat the same search for the previous digit, and so on.

Early versions of that got the right answer, but still were pretty slow
because I hadn't identified a good way to bound the search space for z
and I was searching unnecessarily large ranges.

After solving a few digits with that slow implementation, I stepped
through the program manually with those digits and that's when I
realized the way the coefficients map to when z increases or decreases.
That allowed me to tighten up the bounds for z considerably, and the
search became quite fast.

I never did fully figure out what the checksum algorithm really
*was* myself. But reading through the discussions on reddit, this
explanation was what finally made sense to me:

- https://www.reddit.com/r/adventofcode/comments/rnl0vn/2021_day_24_had_me_for_a_sec/hpu6phh/
- https://github.com/dphilipson/advent-of-code-2021/blob/master/src/days/day24.rs