wip: apply learnings from other days
Using djikstra to pre-calc all the routes as edges with costs & required
keys helps with making the search both readable and performant.

Not performant enough: the more complex example takes over 40s and I
didn't let it finish on my actual input.

The ultimate difficulty of A* for this problem is still a decent f score
heuristic (which I completely dropped for now in the interests of
complexity while I figure this out).

Next step: probably djikstra all the way, pre-filtering possible next
states, treat the whole thing as a graph traversal.