2019 d18
On par with 2018 d23 (octrees) for just totally messing with my brain.
At first glance it seems like a pretty straightforward path-finding
problem you could throw A*, but the size of the search space (26!
That's a factorial, I'm not excited.) combined with the lack of a clear
f score ("distance to goal") heuristic makes it very tricky.

I iterated through several approaches here (still visible on the
day-specific branch), slowly getting at something that was fast enough.
One of the first things I realized was the top-level search shouldn't be
moving one square at a time, but between keys, with key -> key paths a
sub-problem to solve separately.

I had an iteration that did that naively and poorly a while ago. After
my recent experience with 2021 d23, an obvious increment was to use
djikstra to pre-calculate all the optimal paths from point -> point as
part of the maze itself, and then just check against those paths when
finding neighboring states. That key insight, combined with some other
improvements to the search impl (basically a modified djikstra at this
point) and trying to optimize a few of the low-level bits of logic got
this into a pretty good spot.

p1 ran in about 900ms. The easiest way to do p2 was to generalize the
existing code to handle multiple current positions. This did make p1 on
its own slightly slower (1.1s), but it's not too bad. p1 & p2 together
are about 7s. Not amazing, but I'll take it, and I suspect considerably
speeding it up would require making it less readable.