2019 d20 p1
This was all pretty straightforward (although the parsing had some
finicky bits) until I got to the djikstra implementation and remembered
I needed to keep my heap sorted based on a measure not intrinsic to the
thing in the heap (i.e. `dist[pos]`), and that's... hard, with Rust's
borrow checker.

I initially wanted to give my HeapWrapper struct a borrow on &dist,
but that got into needing manual lifetime annotation and that's the part
of Rust that I remain completely befuddled by and unable to grok or fix.

I don't love inserting the cost statically on the HeapWrapper. I guess
because it makes the data redundant with `dist`. Not sure why that
bothers me: I've used similar heap impls in Ruby before, and I often do
implement the states as their own type that can measure their own cost
as well, and with those the dist map is also redundant.

So I guess this is fine?

I'm actually now also realizing/wondering if some of my other
implementations have been subtly broken if they were using dist for
the heap comparison. If I add state x to the queue with cost y based on
dist[x], and later find my way to state x via a different path with cost
z < y, and update dist[x]: the entry of x already in the queue is now in
the wrong spot. That may cause queue.push to put the *new* copy of x in
the wrong spot! Rust is pretty clear this would be a logic error that
violates an invariant of BinaryHeap. cppreference isn't clear about this
with priority_queue or the *_heap functions, but based on my
understanding of the usual way heaps are implemented on top of
vectors, this certainly feels like a potentially subtle and bad bug I've
been implementing lots of places!

p2 is different enough it's going to take me a bit of effort, so I'll
come back to that later.