2019 d23 p2 wip
This seems like the right general direction.

I think the overall cycle count thing was a red herring and not
necessary. sending the nat packet after 1 cycle or 3 doesn't seem to
affect behavior, including what's logged. It's always a pattern  of
sending the same nat packet 3 times in a row, then some work, then a new
nat packet and new idle period, send nat packet 3 times, etc.

I'm getting the same answer as p1 based on that (it sends that packet 3
times), but that can't be right. The problem is pretty clear it's
repetition of *delivering the packet to computer 0*, not repetition of
*a computer sending a packet to the nat*.

p1 and p2 never have the same answer, so there's clearly a bug
somewhere, but I'm not seeing it right now.