2019 d23
p1 was pretty straightforward.

p2 took a bit of massaging to get the definition of "idle" correct.

My first take was just to check at the end of the loop if all machines
were waiting for input or just had the filler in their queue. That was
wrong because when a machine was waiting for input I pushed -1 as input
and then immediately stepped it. So of course it was no longer waiting
for input and no longer had -1 in its input queue.

Counting a machine as becoming idle when it gets -1 and becoming active
when it outputs a packet or receives a packet worked well. I think it
didn't work when I only did it for output - receiving a packet was also
important.

Also: my first time reaching for RefCell in rust. I think this is the
right way to mutate items in a vector while you're iterating over
it?