2022 d04: convert panics to results
Just playing since I don't ever really get to write "real" Rust code
where I need to worry about errors and not just crash at the first sign
of trouble.

Still have 1 expect in `main`, which seems reasonable since it's pretty
equivalent to doing a `match` and logging the error there.

And 1 `unwrap` when parsing. I'm not sure what the appropriate pattern
here is, and  there may not be an elegant idiomatic solution. It looks
like I could `collect` a vec of results into a result of a vec[1] (so
like `let lines: Vec<&str, _> = br.lines().collect()?`), but the whole
reason I've been using `BufReader` is to avoid effectively mapping the
whole file into memory at once (not that it really matters at this
scale, I'm just being picky), so at that point I might as well just use
`fs::read_to_string`. I think if I wanted to maintain the more optimal
memory usage of `BufReader` and convert this to return a result rather
than panic I'd probably just need to do a for loop that appends each
element to a `mut Vec<Rucksack>` as it goes, and then return the result.
It'd be nice if `?` inside a closure could "hoist" to the outer
function, but I can see why they don't do that - that would be skirting
pretty close to behaving like exceptions.

I don't think I'll stick with this for later days. It's a useful
exercise, but it does lead to slightly more complex code and for AOC the
benefit isn't worth it.

[1]: https://stackoverflow.com/questions/63798662/how-do-i-convert-a-vecresultt-e-to-resultvect-e