2022 d07
This was a bit of a nightmare in rust, due to the self-referential
nature of a FS tree. There's probably a much more elegant way to
represent this, and I probably should have just used ruby today, but I
just threw `Rc` and `RefCell` at the problem until it went away.