2022 refactor makefiles for "running all"
I've been keeping this top-level Makefile to quickly get a read on how
long it takes to calculate all solutions. `cargo run` has some overhead,
though, so I'd prefer to just run the existing release binary if it's
there. Tracking all that from the top-level Makefile is a bit of a pain,
so I'm shifting this to day-specific makefiles with the top-level just
using that.