2022 d07 perf improvement experiment
After having an early d08 impl that was slower than necessary because it
kept re-calculating scenic scores for the same points I realized I was
kind of doing the same thing for this one with tot_size for directories.

Did a quick experiment to try memoizing this value, but it didn't speed
things up, and might have even been very slightly slower, so I'm not
gonna bother more. The version I finished yesterday runs in ~5ms even
with the extra work, so not worth worrying about too much.