Remove most input files
Recent reddit discussion: https://www.reddit.com/r/adventofcode/comments/zh2hk0/2022friendly_reminder_dont_commit_your_input/

I had always thought the request for this was just "don't collect many
inputs in one public place for people", but a comment on that thread[1]
does cite a tweet by Eric making it clearer that he would prefer people
just not publish their inputs in general. Doesn't seem like he feels
*that* strongly, and seems pretty fine with people sharing inputs a bit
on places like the subreddit to help people work through their bugs, but
I want to be respectful of his wishes so I'm removing mine.

I am, however, opting to disregard the request on the genre of puzzle
where human interpretation of the input is necessary. I.e. a few times
there have been puzzles where the input is some wildly inefficient
assembler program and the way to get the answer is to figure out what
the program is actually doing and then implement that logic efficiently.

I think half the fun of AOC is the community sharing their approaches &
solution approaches. For most puzzles, reading the puzzle & somebody's
code is enough to grok that, but for these kinds of puzzles the critical
step is the human reading the assembler and writing down what it's
doing, and so not sharing the input with the notes on what it's doing is
hiding the only part of the solution that's really worth sharing IMHO.
This is only a handful of puzzles, and as mentioned above it doesn't
seem like Eric's stance on this is too aggressive, so hopefully this is
a reasonable position to take.

I am also not going to the trouble of rewriting history here: if anyone
stumbles across this repo and reads the history, they'll find the
inputs. Again, my read of Eric's requests is to try to avoid making a
bunch of inputs easily discoverable, so I think just removing them
moving forward does that sufficiently.

[1]: https://www.reddit.com/r/adventofcode/comments/zh2hk0/2022friendly_reminder_dont_commit_your_input/izka2wk/